+++ prisma/migrations/20260511175144_add_venue_image_urls/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "EventVenue" ADD COLUMN     "imageUrls" TEXT[] DEFAULT ARRAY[]::TEXT[];